% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDecimal.R
\name{CreateDecimal}
\alias{CreateDecimal}
\title{CreateDecimal
takes}
\usage{
CreateDecimal(x, y)
}
\arguments{
\item{y}{}
}
\value{
z
}
\description{
CreateDecimal
takes
}
\examples{
x <- c(1,2,3,4,5)
y <- c(2,3,4,3,5)


}
