% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecimaltoPercent.R
\name{DecimaltoPercent}
\alias{DecimaltoPercent}
\title{DecimaltoPercent}
\usage{
DecimaltoPercent(decimal)
}
\arguments{
\item{decimal}{}
}
\value{
percent
}
\description{
DecimaltoPercent
}
\examples{
decEX <- c(0.1, 0.2, 0.4, 0.5)
DecimaltoPercent(decEX)
}
