% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PercenttoDecimal.R
\name{PercenttoDecimal}
\alias{PercenttoDecimal}
\title{PercenttoDecimal}
\usage{
PercenttoDecimal(percent)
}
\arguments{
\item{percent}{}
}
\value{
decimal
}
\description{
PercenttoDecimal
}
\examples{
percEX <- c("15\%", "25\%")
PercenttoDecimal(percEX)
}
